Attempt to fix SVG floor plans.
Accessing the dimensions of an SVG image appears to result in undefined behavior that is handled differently depending on the browser.

In Chrome, an attempt is made to determine the intrinsic dimensions of an image (for citec-gebaeudeplan.svg, this value is 150 x 150). However, Firefox sets the dimensions of SVG images to 0 x 0.

The problem arises when the getBounds method (located in FloorPlan.vue:36) divides by the width/height of the image, leading to an attempt to create coordinates with NaN values, which results in an error.

To mitigate this issue, I've implemented a temporary solution by manually setting the width and height to 150px if the image element's width and height are both 0. It's important to note that this solution is not ideal and may not work correctly for non-square SVG floor plans.